# Conflicts:
#	app/build.gradle
#	app/src/main/java/com/imcys/bilibilias/home/ui/fragment/HomeFragment.kt
#	common/src/main/java/com/imcys/bilibilias/common/base/AbsActivity.kt

+++ app/src/main/java/com/imcys/bilibilias/base/model/user/DownloadTaskBean.kt
@@ -8,12 +8,12 @@ import com.imcys.bilibilias.home.ui.model.VideoPageListData
 import com.imcys.bilibilias.home.ui.model.VideoPlayBean
 
 data class DownloadTaskDataBean(
-    val cid: Int,
+    val cid: Long,
     val pageTitle: String,
     val bvid: String,
-    //分辨率
+    // 分辨率
     val qn: String,
-    //视频获取方式选择
+    // 视频获取方式选择
     val fnval: String = "80",
     //
     val platform: String = "pc",
@@ -24,4 +24,4 @@ data class DownloadTaskDataBean(
     val videoPageDataData: VideoPageListData.DataBean? = null,
     val bangumiSeasonBean: BangumiSeasonBean.ResultBean.EpisodesBean? = null,
 
-    )
+)
